requirements.txt

# Core dependencies
python-dotenv==1.0.0
numpy==1.24.3
faiss-cpu==1.7.4
requests==2.31.0
langchain-core==0.1.6
langchain-groq==0.0.1
langchain-openai==0.0.3
langchain-anthropic==0.1.2
langchain-google-genai==0.0.5
langchain-ollama==0.0.1

# File processing
nbformat==5.9.2
pypdf==4.0.0
python-docx==1.0.1
openpyxl==3.1.2
python-pptx==0.6.22
pyarrow==14.0.1
beautifulsoup4==4.12.2
pyyaml==6.0.1
h5py==3.10.0
lxml==4.9.3

# Web UI
flask==2.3.3
werkzeug==2.3.7
